We have tests around a Firebase database wrapper

+++ core/src/main/java/com/novoda/materialised/hackernews/database/TopStories.kt
@@ -1,5 +1,5 @@
 package com.novoda.materialised.hackernews.database
 
 interface TopStories {
-    fun readAll(callback: ValueCallback<List<Int>>)
+    fun readAll(callback: ValueCallback<List<Long>>)
 }
